fix: component breakdown and inline styling to emotion css
