+++ package.json
@@ -14,6 +14,7 @@
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
     "@types/react-router-dom": "^5.1.6",
+    "emotion": "^10.0.27",
     "formik": "^2.2.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
