+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import Dashboard from "./components/dashboard/Dashboard";
+import Dashboard from "./components/Dashboard";
 
 function App() {
   return (
