+++ src/components/Dashboard/components/GamesColumn/styles.ts
@@ -1,0 +1,32 @@
+import { css } from "emotion";
+
+export const appBar = css`
+  box-shadow: none !important;
+`;
+
+export const paper = css`
+  width: 100%;
+  box-shadow: none !important;
+`;
+
+export const tableContainer = css`
+  height: 450px;
+`;
+
+export const tableCell = css`
+  cursor: pointer;
+`;
+
+export const noGamesDiv = css`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const createButton = css`
+  border-radius: 4px !important;
+  margin-top: 10px !important;
+`;
