+++ src/components/Dashboard/components/OptionsColumn/styles.ts
@@ -1,0 +1,6 @@
+import { css } from "emotion";
+
+export const button = css`
+  width: 200px;
+  borderradius: 4;
+`;
