+++ src/components/Dashboard/styles.ts
@@ -1,0 +1,12 @@
+import { css } from "emotion";
+
+export const appBar = css`
+  box-shadow: none !important;
+`;
+
+export const gridItem = css`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
