+++ src/components/Navbar/styles.ts
@@ -1,0 +1,23 @@
+import { css } from "emotion";
+
+export const appBar = css`
+  box-shadow: none !important;
+  margin-top: -8px;
+`;
+
+export const toolBar = css`
+  padding: 0px 0px 0px 16px !important;
+`;
+
+export const appNameGrid = css`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+`;
+
+export const signInButton = css`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
